integrate capture_glob into ConventionResolver
