--- hydromt/utils/constants.py
@@ -1,4 +1,0 @@
-"""Publically available constants for hydromt."""
-
-DEFAULT_TABLE_FILENAME = "tables/{name}.csv"
-DEFAULT_GEOM_FILENAME = "geoms/{name}.geojson"
